// src/components/types.ts
export enum FilterByPrice {
  // MostRecent = "Most Recent",
  LowestPrice = "Lowest Price",
  HighestPrice = "Highest Price",
}

export enum FilterByCategory {
  All = "All",
  Laptops = "Laptops",
  Cameras = "Cameras",
  Phones = "Phones",
  TabletsAndEReaders = "Tablets & E-Readers",
  PhoneAccessories = "Phone Accessories",
  PCAccessories = "PC Accessories",
  Gaming = "Gaming",
  Audio = "Audio",
  SmartHome = "Smart Home",
  MonitorsAndTv = "Monitors & TV",
  Drones = "Drones",
}